Rename project from Speed Dreams to speed-dreams
Since the generated name of the AppImage is name-version-arch.AppImage,
the space in "Speed Dreams" was the only space in the filename, while
the rest of the name was separated using dashes and underscores.
This change removes the space from the name of the AppImage, meaning
there are no spaces left in the AppImage's filenmae either.